feat: 테마함 navigation 추가

+++ app/src/main/java/com/umc/halo/presentation/storybook/detail/StoryBookDetailUiState.kt
@@ -27,14 +27,14 @@ data class StoryBookInfo(
 )
 
 data class StoryBookIndex(
-    val id: Int,
+    val id: Long,
     val title: String,
     val subTitle: String,
     val isLocked: Boolean
 )
 
 data class TodayStoryBook(
-    val id: Int,
+    val id: Long,
     val title: String,
     val tag: String,
     val isLocked: Boolean
